requirements.txt
torch>=2.2.0
torchvision>=0.15.0
streamlit==1.25.0
numpy==1.23.0
pillow==9.5.0
scikit-learn==1.3.0
